requirements.txt
requests
urllib3
pymysql
sqlalchemy
lxml
beautifulsoup4
python-dotenv
